+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ### Bugs Fixed
 
 - Fixed best practice tool invocation failure when passing "all" action with "general" or "azurefunctions" resources, by adding that support. [[#757](https://github.com/Azure/azure-mcp/issues/757)]
+- Update CREATE and SET operations to destructive = true. [[#773](https://github.com/Azure/azure-mcp/pull/773)]
 
 ### Other Changes